reorganize utils, move some to "common" package
