+++ app/components/github_integration/code_links.py
@@ -10,17 +10,17 @@ from typing import NamedTuple
 import discord
 from zig_codeblocks import highlight_zig_code
 
-from app.components.zig_codeblocks import THEME
-from app.setup import gh
-from app.utils import (
+from app.common.cache import TTRCache
+from app.common.hooks import (
     ItemActions,
     MessageLinker,
-    TTRCache,
+    ProcessedMessage,
     create_delete_hook,
     create_edit_hook,
     remove_view_after_timeout,
 )
-from app.utils.hooks import ProcessedMessage
+from app.components.zig_codeblocks import THEME
+from app.setup import gh
 
 CODE_LINK_PATTERN = re.compile(
     r"https?://(?:www\.)?github\.com/([^/\s]+)/([^/\s]+)/blob/([^/\s]+)/([^\?#\s]+)"
