+++ app/components/github_integration/mentions/integration.py
@@ -1,13 +1,15 @@
 import discord
 
 from .fmt import entity_message
-from app.utils import (
+from app.common.hooks import (
     ItemActions,
     MessageLinker,
     create_delete_hook,
     create_edit_hook,
-    is_dm,
     remove_view_after_timeout,
+)
+from app.utils import (
+    is_dm,
     try_dm,
 )
